Refactor(jungyonghoon/develop): 폴더 구조 최적화
