+++ mohae/src/app.module.ts
@@ -2,7 +2,7 @@ import { CacheInterceptor, CacheModule, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { typeORMConfig } from './configs/typeorm.config';
+import { typeORMConfig } from './common/configs/typeorm.config';
 import { ReportsModule } from './reports/reports.module';
 import { FaqsModule } from './faqs/faqs.module';
 import { CategoriesModule } from './categories/categories.module';
