+++ mohae/src/auth/auth.module.ts
@@ -14,8 +14,8 @@ import { MajorsModule } from 'src/majors/majors.module';
 import { JwtStrategy } from './jwt/jwt.strategy';
 import { CategoryRepository } from 'src/categories/repository/category.repository';
 import { CategoriesModule } from 'src/categories/categories.module';
-import { ErrorConfirm } from 'src/utils/error';
-import { LoginProcess } from 'src/utils/login';
+import { ErrorConfirm } from 'src/common/utils/error';
+import { LoginProcess } from 'src/common/utils/login';
 import { User } from './entity/user.entity';
 import { SpecRepository } from 'src/specs/repository/spec.repository';
 import { LikeRepository } from 'src/like/repository/like.repository';
