+++ mohae/src/faqs/faqs.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserRepository } from 'src/auth/repository/user.repository';
-import { ErrorConfirm } from 'src/utils/error';
+import { ErrorConfirm } from 'src/common/utils/error';
 import { FaqsController } from './faqs.controller';
 import { FaqsService } from './faqs.service';
 import { FaqRepository } from './repository/faq.repository';
