+++ mohae/src/reports/reports.controller.spec.ts
@@ -2,7 +2,7 @@ import { PassportModule } from '@nestjs/passport';
 import { Test, TestingModule } from '@nestjs/testing';
 import { UserRepository } from 'src/auth/repository/user.repository';
 import { BoardRepository } from 'src/boards/repository/board.repository';
-import { ErrorConfirm } from 'src/utils/error';
+import { ErrorConfirm } from 'src/common/utils/error';
 import { CreateReportDto } from './dto/report.dto';
 import { ReportsController } from './reports.controller';
 import { ReportsService } from './reports.service';
