+++ mohae/src/reviews/reviews.module.ts
@@ -8,7 +8,7 @@ import { BoardsService } from 'src/boards/boards.service';
 import { BoardRepository } from 'src/boards/repository/board.repository';
 import { CategoryRepository } from 'src/categories/repository/category.repository';
 import { BoardPhotoRepository } from 'src/photo/repository/photo.repository';
-import { ErrorConfirm } from 'src/utils/error';
+import { ErrorConfirm } from 'src/common/utils/error';
 import { ReviewRepository } from './repository/review.repository';
 import { ReviewsController } from './reviews.controller';
 import { ReviewsService } from './reviews.service';
