+++ mohae/src/reviews/reviews.service.ts
@@ -2,7 +2,7 @@ import { Injectable, InternalServerErrorException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UserRepository } from 'src/auth/repository/user.repository';
 import { BoardRepository } from 'src/boards/repository/board.repository';
-import { ErrorConfirm } from 'src/utils/error';
+import { ErrorConfirm } from 'src/common/utils/error';
 import { CreateReviewDto } from './dto/review.dto';
 import { Review } from './entity/review.entity';
 import { ReviewRepository } from './repository/review.repository';
