+++ mohae/src/specs/specs.module.ts
@@ -4,7 +4,7 @@ import { AuthModule } from 'src/auth/auth.module';
 import { UserRepository } from 'src/auth/repository/user.repository';
 import { PhotoModule } from 'src/photo/photo.module';
 import { SpecPhotoRepository } from 'src/photo/repository/photo.repository';
-import { ErrorConfirm } from 'src/utils/error';
+import { ErrorConfirm } from 'src/common/utils/error';
 import { SpecRepository } from './repository/spec.repository';
 import { SpecsController } from './specs.controller';
 import { SpecsService } from './specs.service';
